python3 changes

+++ src/utils/utils.py
@@ -115,16 +115,18 @@ def checkboard(I1, I2, n=7):
     assert I1.shape == I2.shape
     height, width, channels = I1.shape
     hi, wi = height/n, width/n
-    outshape = (hi*n, wi*n, channels)
+    outshape = (int(hi*n), int(wi*n), int(channels))
+
 
     out_image = np.zeros(outshape, dtype='uint8')
     for i in range(n):
-        h = hi * i
-        h1 = h + hi
+        h = int(hi * i)
+        h1 = int(h + hi)
         for j in range(n):
-            w = wi * j
-            w1 = w + wi
+            w = int(wi * j)
+            w1 = int(w + wi)
             if (i-j)%2 == 0:
+
                 out_image[h:h1, w:w1, :] = I1[h:h1, w:w1, :]
             else:
                 out_image[h:h1, w:w1, :] = I2[h:h1, w:w1, :]
